Fix for linuxmint
Before this patch, ncurses would freeze after libgame.so was
reloaded in main.c. After posting the issue on Stackoverflow,
it's because the game shared lib owns ncurses due to the
--as-needed switch to the linker being default in this
(distro|version of GCC).

The fix recommended by the Stackoverflow answer causes the
display to drop back to the terminal for an instant, so I
changed it a little.

main.c now starts up and brings down ncurses, so it now owns
the all important handle to the screen, . Now when
the game library is recompiled, this screen handle no longer
goes out of scope, and the game continues as expected!

* http://stackoverflow.com/questions/27881111